Add content to support arrow link componentisation
This change is to support the componentisation of the highlight arrow
link that is used in a few different variants across the corona virus
content pages.

Currently we are using an nbsp character to prevent the last word
wrapping to the next line on smaller screens for the page header version
of this element.

As part of our work to componentise the highlight/arrow link, we have
adopted a different approach to preventing undesired word wrapping for
the component.

The text containing an nbsp character will become redundant and is to be
removed once the componentisation work has safely made it to the live
site.