Check links are OK in continuous integration
In the past we've accidentally merged links with typos in them, for
example missing a trailing character (like the one fixed in #201).

We could fix this by checking that all the links in the YAML respond
with 200 OK status codes.

It's possible that this check might be a bit too strict, as it will also
point out redirects, which we probably don't want, but they're not
really the end of the world.

It's also possible that this could be annoying when we're adding links
to content that's still being drafted but hasn't been published, as we
might want to do something like:

1) Draft some new content (but don't publish it)
2) Update this content to point at the drafted link
3) Get that merged, and deployed as a draft
4) Check it's all good in draft
5) Publish the new content
6) Publish this content (now that the link isn't broken)

We also might need to teach the content designers a bit more about
GitHub to help them diagnose what's wrong when the build fails.
While we've still got developers working that's less of an issue, but it
would be nice not to have to rely on that forever.