Switch tests to use rspec
Currently the "tests" only use the standard library. This is fine, but
it might be a bit more familiar to use the rspec library, since we use
that everywhere else on GOV.UK.

This is partly motivated by #203 (which it will conflict with), in which
I ended up re-implementing bits of rspec's output.