Use rspec for link tests
Previously we didn't have rspec in this repo, so the tests were written
using the standard library. We can save ourselves a bit of gross work in
formatting the output by using rspec instead.

Running this with `bundle exec rspec --format documentation` gives
output like this:

```
coronavirus_education_page.yml
  /content/sections[3]/sub_sections[0]/list[1]/url
    should not have a broken link to `/government/publications/coronavirus-covid-19-financial-support-for-education-early-years-and-childrens-social-care`
  /content/sections[2]/sub_sections[0]/list[3]/url
    should not have a broken link to `/government/publications/coronavirus-covid-19-early-years-and-childcare-closures`
  /content/sections[4]/sub_sections[0]/list[1]/url
    should not have a broken link to `/guidance/guidance-for-prospective-students`
  /content/sections[1]/sub_sections[0]/list[6]/url
    should not have a broken link to `/government/publications/coronavirus-covid-19-send-risk-assessment-guidance`
```

When a test fails, it will output:

* The HTTP status code it got (e.g. 404)
* The link it was testing
* The file it found the link in
* The path to the bit of the file that path is in

Which should be more than enough to work out where the issue is.