Reuse TCP connections for links to the same host
In the previous commit, each test would establish a fresh TCP connection
to each link, which is a bit slow and puts undue load on the poor
servers.

We can reuse TCP connections by grouping links by host, which reduces
the time the tests take to run from ~30 seconds to ~10 seconds.

This also allows us to specify a longer timeout for setting up the
connection, which might help with the tests that are failing because
they can't connect to gov.wales within the timeout.

The tests now assume that all links are https on port 443, which is true
for now but could change in the future in theory.