Update coronavirus_landing_page.yml
1. Deleted line 54 ('Do not leave home for 10 days after your test').
Because: It wasn't displaying. Think this is down to wrong code being used - a dev will take a look tomorrow, but in the meantime I need to do a workaround...
2. Added '- do not leave home for at least 10 days after your test' to the link test (line 53).
Because: It's the best place to put it until we can find a way to display it on its own line.