Update coronavirus_business_page.yml
1.  CHANGED line 9 FROM
Employers can apply for staff to get up to 80% pay if they can’t work
TO
Employers may be eligible for financial support to pay wages
BECAUSE 
New scheme is coming in from 1 Nov with different percentages. The old scheme will sort of be running alongside it for a bit. (Hence this line is intentionally generic for now.)

2. 'What you can do now ' section
CHANGED order of arrowed bits, following feedback from Leonie/Checan   and an analytics check.